fix: address remaining coderabbit feedback on PR #544
- Extract ConceptGraph to concept_graph.rs (keep mod.rs under 500 LOC)

- edge.rs: add confidence clamping with clamp_confidence_val + custom serde deserializer

- node.rs: clarify add_episode_ref docstring (not counting as access), document +2 heuristic in estimated_tokens

- state.rs: optimize get_shared_context with HashMap frequency counting
