+++ memory-core/src/retrieval/cascade/mod.rs
@@ -9,7 +9,10 @@
 //! The cascade eliminates 50-70% of embedding API calls by satisfying
 //! queries from CPU-local tiers before falling back to the API.
 
-use anyhow::{Context, Result};
+use anyhow::Result;
+
+mod concept_graph;
+pub use concept_graph::ConceptGraph;
 
 /// Configuration for the cascading retrieval pipeline.
 #[derive(Debug, Clone)]
@@ -92,115 +95,6 @@ pub struct CascadeResult {
     pub contributing_tiers: Vec<String>,
     /// Number of API calls made (should be 0 or 1).
     pub api_calls: u32,
-}
-
-/// A simple concept graph ontology entry for domain-term expansion.
-#[derive(Debug, Clone)]
-struct OntologyEntry {
-    #[allow(dead_code)] // domain name may be useful for diagnostics
-    domain: String,
-    terms: Vec<String>,
-    related_concepts: Vec<String>,
-}
-
-/// In-memory concept graph for query term expansion (Tier 3).
-///
-/// Loaded from the embedded `ontology.json` at compile time.
-#[derive(Debug, Clone)]
-pub struct ConceptGraph {
-    entries: Vec<OntologyEntry>,
-}
-
-impl ConceptGraph {
-    /// Create a new ConceptGraph from the embedded ontology JSON.
-    #[must_use]
-    pub fn from_embedded() -> Self {
-        let json = include_str!("ontology.json");
-        Self::from_json(json).unwrap_or_else(|e| {
-            tracing::warn!(
-                error = %e,
-                "Failed to parse embedded ontology, using empty concept graph"
-            );
-            Self {
-                entries: Vec::new(),
-            }
-        })
-    }
-
-    /// Parse a ConceptGraph from JSON.
-    fn from_json(json: &str) -> Result<Self> {
-        let parsed: serde_json::Value =
-            serde_json::from_str(json).context("Failed to parse ontology JSON")?;
-
-        let domains = parsed["domains"]
-            .as_array()
-            .context("Missing 'domains' array in ontology")?;
-
-        let entries: Vec<OntologyEntry> = domains
-            .iter()
-            .map(|d| OntologyEntry {
-                domain: d["domain"].as_str().unwrap_or("").to_string(),
-                terms: d["terms"]
-                    .as_array()
-                    .map(|a| {
-                        a.iter()
-                            .filter_map(|v| v.as_str().map(String::from))
-                            .collect()
-                    })
-                    .unwrap_or_default(),
-                related_concepts: d["related_concepts"]
-                    .as_array()
-                    .map(|a| {
-                        a.iter()
-                            .filter_map(|v| v.as_str().map(String::from))
-                            .collect()
-                    })
-                    .unwrap_or_default(),
-            })
-            .collect();
-
-        Ok(Self { entries })
-    }
-
-    /// Expand query terms using the ontology.
-    ///
-    /// For each word in the query, finds matching domain terms and returns
-    /// all synonyms and related concepts.
-    #[must_use]
-    pub fn expand_terms(&self, query: &str) -> Vec<String> {
-        let query_lower = query.to_lowercase();
-        let query_words: Vec<&str> = query_lower.split_whitespace().collect();
-
-        let mut expanded: Vec<String> = Vec::new();
-
-        for entry in &self.entries {
-            let domain_matched = query_words
-                .iter()
-                .any(|w| entry.terms.iter().any(|t| t == w));
-
-            if domain_matched {
-                // Add all terms and related concepts from matching domains
-                expanded.extend(entry.terms.iter().cloned());
-                expanded.extend(entry.related_concepts.iter().cloned());
-            }
-        }
-
-        expanded.sort();
-        expanded.dedup();
-        expanded
-    }
-
-    /// Check if the concept graph has any entries.
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.entries.is_empty()
-    }
-
-    /// Get the number of ontology domains.
-    #[must_use]
-    pub fn domain_count(&self) -> usize {
-        self.entries.len()
-    }
 }
 
 /// Cascading retrieval orchestrator.
